+++ packages/access-control/lib/access-control.js
@@ -1,4 +1,4 @@
-const { getType, pick, defaultObj } = require('@keystone-alpha/utils');
+const { getType, pick, defaultObj, intersection } = require('@keystone-alpha/utils');
 
 const validateGranularConfigTypes = (longHandAccess, validationError) => {
   const errors = Object.entries(longHandAccess)
@@ -8,40 +8,10 @@ const validateGranularConfigTypes = (longHandAccess, validationError) => {
   if (errors.length) {
     throw new Error(errors.join('\n'));
   }
+  return longHandAccess;
 };
 
-const parseGranularAccessConfig = ({
-  accessTypes,
-  access,
-  defaultAccess,
-  onGranularParseError,
-  validateGranularType,
-}) => {
-  const longHandAccess = pick(access, accessTypes);
-
-  // An object was supplied, but it has the wrong keys (it's probably a
-  // declarative access control config being used as a shorthand, which
-  // isn't possible [due to `create` not supporting declarative config])
-  if (Object.keys(longHandAccess).length === 0) {
-    onGranularParseError();
-  }
-  // Construct an object with all keys
-  const finalAccess = {
-    ...defaultObj(accessTypes, defaultAccess),
-    ...longHandAccess,
-  };
-  validateGranularConfigTypes(finalAccess, validateGranularType);
-
-  return finalAccess;
-};
-
-const parseAccess = ({
-  accessTypes,
-  access,
-  defaultAccess,
-  onGranularParseError,
-  validateGranularType,
-}) => {
+const parseAccessCore = ({ accessTypes, access, defaultAccess, onGranularParseError }) => {
   const type = getType(access);
   switch (type) {
     case 'Boolean':
@@ -49,13 +19,13 @@ const parseAccess = ({
       return defaultObj(accessTypes, access);
 
     case 'Object':
-      return parseGranularAccessConfig({
-        accessTypes,
-        access,
-        defaultAccess,
-        onGranularParseError,
-        validateGranularType,
-      });
+      // An object was supplied, but it has the wrong keys (it's probably a
+      // declarative access control config being used as a shorthand, which
+      // isn't possible [due to `create` not supporting declarative config])
+      if (Object.keys(pick(access, accessTypes)).length === 0) {
+        onGranularParseError();
+      }
+      return { ...defaultObj(accessTypes, defaultAccess), ...pick(access, accessTypes) };
 
     default:
       throw new Error(
@@ -64,59 +34,139 @@ const parseAccess = ({
   }
 };
 
-module.exports = {
-  parseListAccess({ listKey, defaultAccess, access = defaultAccess }) {
-    const accessTypes = ['create', 'read', 'update', 'delete'];
+const parseAccess = ({ schemaNames, accessTypes, access, defaultAccess, parseAndValidate }) => {
+  // Check that none of the schemaNames match the accessTypes
+  if (intersection(schemaNames, accessTypes).length > 0) {
+    throw new Error(
+      `${JSON.stringify(
+        intersection(schemaNames, accessTypes)
+      )} are reserved words and cannot be used as schema names.`
+    );
+  }
 
-    return parseAccess({
-      accessTypes,
-      access,
-      defaultAccess,
-      onGranularParseError: () => {
+  const providedNameCount = intersection(Object.keys(access), schemaNames).length;
+  const type = getType(access);
+
+  if (
+    type === 'Object' &&
+    providedNameCount > 0 &&
+    providedNameCount < Object.keys(access).length
+  ) {
+    // If some are in, and some are out, throw an error!
+    throw new Error(
+      `Invalid schema names: ${JSON.stringify(
+        Object.keys(access).filter(k => !schemaNames.includes(k))
+      )}`
+    );
+  }
+
+  const namesProvided = type === 'Object' && providedNameCount === Object.keys(access).length;
+  return schemaNames.reduce(
+    (acc, schemaName) => ({
+      ...acc,
+      [schemaName]: parseAndValidate(
+        namesProvided
+          ? access.hasOwnProperty(schemaName) // If all the keys are in schemaNames, parse each on their own
+            ? access[schemaName]
+            : defaultAccess
+          : access
+      ),
+    }),
+    {}
+  );
+};
+
+module.exports = {
+  parseCustomAccess({ defaultAccess, access = defaultAccess, schemaNames }) {
+    const accessTypes = [];
+    const parseAndValidate = access => {
+      const type = getType(access);
+      if (!['Boolean', 'Function', 'Object'].includes(type)) {
         throw new Error(
-          `Must specify one of ${JSON.stringify(
-            accessTypes
-          )} access configs, but got ${JSON.stringify(
-            Object.keys(access)
-          )}. (Did you mean to specify a declarative access control config? This can be done on a granular basis only)`
+          `Expected a Boolean, Object, or Function for custom access, but got ${type}`
         );
-      },
-      validateGranularType: (type, accessType) => {
-        if (accessType === 'create') {
-          if (!['Boolean', 'Function'].includes(type)) {
-            return `Expected a Boolean, or Function for ${listKey}.access.${accessType}, but got ${type}. (NOTE: 'create' cannot have a Declarative access control config)`;
-          }
-        } else {
-          if (!['Object', 'Boolean', 'Function'].includes(type)) {
-            return `Expected a Boolean, Object, or Function for ${listKey}.access.${accessType}, but got ${type}`;
-          }
-        }
-      },
-    });
+      }
+      return access;
+    };
+    return parseAccess({ schemaNames, accessTypes, access, defaultAccess, parseAndValidate });
   },
 
-  parseFieldAccess({ listKey, fieldKey, defaultAccess, access = defaultAccess }) {
-    const accessTypes = ['create', 'read', 'update'];
-
-    return parseAccess({
-      accessTypes,
-      access,
-      defaultAccess,
-      onGranularParseError: () => {
-        throw new Error(
-          `Must specify one of ${JSON.stringify(
-            accessTypes
-          )} access configs, but got ${JSON.stringify(
-            Object.keys(access)
-          )}. (Did you mean to specify a declarative access control config? This can be done on lists only)`
-        );
-      },
-      validateGranularType: (type, accessType) => {
-        if (!['Boolean', 'Function'].includes(type)) {
-          return `Expected a Boolean or Function for ${listKey}.fields.${fieldKey}.access.${accessType}, but got ${type}. (NOTE: Fields cannot have declarative access control config)`;
+  parseListAccess({ listKey, defaultAccess, access = defaultAccess, schemaNames }) {
+    const accessTypes = ['create', 'read', 'update', 'delete', 'auth'];
+    const parseAndValidate = access =>
+      validateGranularConfigTypes(
+        parseAccessCore({
+          accessTypes,
+          access,
+          defaultAccess,
+          onGranularParseError: () => {
+            throw new Error(
+              `Must specify one of ${JSON.stringify(
+                accessTypes
+              )} access configs, but got ${JSON.stringify(
+                Object.keys(access)
+              )}. (Did you mean to specify a declarative access control config? This can be done on a granular basis only)`
+            );
+          },
+        }),
+        (type, accessType) => {
+          if (accessType === 'create') {
+            if (!['Boolean', 'Function'].includes(type)) {
+              return `Expected a Boolean, or Function for ${listKey}.access.${accessType}, but got ${type}. (NOTE: 'create' cannot have a Declarative access control config)`;
+            }
+          } else {
+            if (!['Object', 'Boolean', 'Function'].includes(type)) {
+              return `Expected a Boolean, Object, or Function for ${listKey}.access.${accessType}, but got ${type}`;
+            }
+          }
         }
-      },
-    });
+      );
+    return parseAccess({ schemaNames, accessTypes, access, defaultAccess, parseAndValidate });
+  },
+
+  parseFieldAccess({ listKey, fieldKey, defaultAccess, access = defaultAccess, schemaNames }) {
+    const accessTypes = ['create', 'read', 'update'];
+    const parseAndValidate = access =>
+      validateGranularConfigTypes(
+        parseAccessCore({
+          accessTypes,
+          access,
+          defaultAccess,
+          onGranularParseError: () => {
+            throw new Error(
+              `Must specify one of ${JSON.stringify(
+                accessTypes
+              )} access configs, but got ${JSON.stringify(
+                Object.keys(access)
+              )}. (Did you mean to specify a declarative access control config? This can be done on lists only)`
+            );
+          },
+        }),
+        (type, accessType) => {
+          if (!['Boolean', 'Function'].includes(type)) {
+            return `Expected a Boolean or Function for ${listKey}.fields.${fieldKey}.access.${accessType}, but got ${type}. (NOTE: Fields cannot have declarative access control config)`;
+          }
+        }
+      );
+    return parseAccess({ schemaNames, accessTypes, access, defaultAccess, parseAndValidate });
+  },
+
+  validateCustomAccessControl({ access, authentication = {} }) {
+    // Either a boolean or an object describing a where clause
+    let result;
+    if (typeof access !== 'function') {
+      result = access;
+    } else {
+      result = access({ authentication: authentication.item ? authentication : {} });
+    }
+    const type = getType(result);
+
+    if (!['Object', 'Boolean'].includes(type)) {
+      throw new Error(
+        `Must return an Object or Boolean from Imperative or Declarative access control function. Got ${type}`
+      );
+    }
+    return result;
   },
 
   validateListAccessControl({ access, listKey, operation, authentication = {}, originalInput }) {