+++ packages/fields-auto-increment/src/Implementation.js
@@ -20,13 +20,13 @@ export class AutoIncrementImplementation extends Implementation {
       : gqlTypeDefault;
   }
 
-  get gqlOutputFields() {
+  gqlOutputFields() {
     return [`${this.path}: ${this.gqlType}`];
   }
-  get gqlOutputFieldResolvers() {
+  gqlOutputFieldResolvers() {
     return { [`${this.path}`]: item => item[this.path] };
   }
-  get gqlQueryInputFields() {
+  gqlQueryInputFields() {
     return [
       ...this.equalityInputFields(this.gqlType),
       ...this.orderingInputFields(this.gqlType),