+++ packages/keystone/lib/Keystone/index.js
@@ -3,6 +3,8 @@ const fs = require('fs');
 const gql = require('graphql-tag');
 const flattenDeep = require('lodash.flattendeep');
 const fastMemoize = require('fast-memoize');
+const falsey = require('falsey');
+const createCorsMiddleware = require('cors');
 const { print } = require('graphql/language/printer');
 const { graphql } = require('graphql');
 const {
@@ -12,16 +14,20 @@ const {
   objMerge,
   flatten,
   unique,
+  filterValues,
 } = require('@keystone-alpha/utils');
 const {
   validateFieldAccessControl,
   validateListAccessControl,
+  validateCustomAccessControl,
+  parseCustomAccess,
 } = require('@keystone-alpha/access-control');
 const {
   startAuthedSession,
   endAuthedSession,
   commonSessionMiddleware,
 } = require('@keystone-alpha/session');
+const { logger } = require('@keystone-alpha/logger');
 
 const {
   unmergeRelationships,
@@ -30,8 +36,11 @@ const {
 } = require('./relationship-utils');
 const List = require('../List');
 const { DEFAULT_DIST_DIR } = require('../../constants');
+const { AccessDeniedError } = require('../List/graphqlErrors');
 
 const debugGraphQLSchemas = () => !!process.env.DEBUG_GRAPHQL_SCHEMAS;
+
+const graphqlLogger = logger('graphql');
 
 module.exports = class Keystone {
   constructor({
@@ -44,10 +53,13 @@ module.exports = class Keystone {
     onConnect,
     cookieSecret = 'qwerty',
     sessionStore,
+    secureCookies = process.env.NODE_ENV === 'production', // Default to true in production
+    cookieMaxAge = 1000 * 60 * 60 * 24 * 30, // 30 days
+    schemaNames = ['public'],
   }) {
     this.name = name;
     this.adapterConnectOptions = adapterConnectOptions;
-    this.defaultAccess = { list: true, field: true, ...defaultAccess };
+    this.defaultAccess = { list: true, field: true, custom: true, ...defaultAccess };
     this.auth = {};
     this.lists = {};
     this.listsArray = [];
@@ -57,9 +69,12 @@ module.exports = class Keystone {
     this._extendedMutations = [];
     this._graphQLQuery = {};
     this._cookieSecret = cookieSecret;
+    this._secureCookies = secureCookies;
+    this._cookieMaxAge = cookieMaxAge;
     this._sessionStore = sessionStore;
-    this.registeredTypes = new Set();
     this.eventHandlers = { onConnect };
+    this.registeredTypes = new Set();
+    this._schemaNames = schemaNames;
 
     if (adapters) {
       this.adapters = adapters;
@@ -68,7 +83,7 @@ module.exports = class Keystone {
       this.adapters = { [adapter.constructor.name]: adapter };
       this.defaultAdapter = adapter.constructor.name;
     } else {
-      throw new Error('Need an adapter, yo');
+      throw new Error('No database adapter provided');
     }
 
     // Placeholder until keystone.prepare() is run during which this function
@@ -91,20 +106,29 @@ module.exports = class Keystone {
   // The GraphQL App uses this method to build up the context required for each
   // incoming query.
   // It is also used for generating the `keystone.query` method
-  getGraphQlContext({ schemaName = 'admin', req = {}, skipAccessControl = false } = {}) {
+  getGraphQlContext({ schemaName, req = {}, skipAccessControl = false } = {}) {
+    let getCustomAccessControlForUser;
     let getListAccessControlForUser;
     let getFieldAccessControlForUser;
 
     if (skipAccessControl) {
+      getCustomAccessControlForUser = () => true;
       getListAccessControlForUser = () => true;
       getFieldAccessControlForUser = () => true;
     } else {
       // memoizing to avoid requests that hit the same type multiple times.
       // We do it within the request callback so we can resolve it based on the
       // request info ( like who's logged in right now, etc)
+      getCustomAccessControlForUser = fastMemoize(access => {
+        return validateCustomAccessControl({
+          access: access[schemaName],
+          authentication: { item: req.user, listKey: req.authedListKey },
+        });
+      });
+
       getListAccessControlForUser = fastMemoize((listKey, originalInput, operation) => {
         return validateListAccessControl({
-          access: this.lists[listKey].access,
+          access: this.lists[listKey].access[schemaName],
           originalInput,
           operation,
           authentication: { item: req.user, listKey: req.authedListKey },
@@ -115,7 +139,7 @@ module.exports = class Keystone {
       getFieldAccessControlForUser = fastMemoize(
         (listKey, fieldKey, originalInput, existingItem, operation) => {
           return validateFieldAccessControl({
-            access: this.lists[listKey].fieldsByPath[fieldKey].access,
+            access: this.lists[listKey].fieldsByPath[fieldKey].access[schemaName],
             originalInput,
             existingItem,
             operation,
@@ -134,6 +158,7 @@ module.exports = class Keystone {
       endAuthedSession: endAuthedSession.bind(null, req),
       authedItem: req.user,
       authedListKey: req.authedListKey,
+      getCustomAccessControlForUser,
       getListAccessControlForUser,
       getFieldAccessControlForUser,
     };
@@ -170,6 +195,7 @@ module.exports = class Keystone {
       };
 
       if (skipAccessControl) {
+        passThroughContext.getCustomAccessControlForUser = () => true;
         passThroughContext.getListAccessControlForUser = () => true;
         passThroughContext.getFieldAccessControlForUser = () => true;
       }
@@ -221,6 +247,7 @@ module.exports = class Keystone {
         }
         return this.createList(auxKey, auxConfig, { isAuxList: true });
       },
+      schemaNames: this._schemaNames,
     });
     this.lists[key] = list;
     this.listsArray.push(list);
@@ -229,9 +256,18 @@ module.exports = class Keystone {
   }
 
   extendGraphQLSchema({ types = [], queries = [], mutations = [] }) {
-    this._extendedTypes = this._extendedTypes.concat(types);
-    this._extendedQueries = this._extendedQueries.concat(queries);
-    this._extendedMutations = this._extendedMutations.concat(mutations);
+    const _parseAccess = obj => ({
+      ...obj,
+      access: parseCustomAccess({
+        access: obj.access,
+        schemaNames: this._schemaNames,
+        defaultAccess: this.defaultAccess.custom,
+      }),
+    });
+
+    this._extendedTypes = this._extendedTypes.concat(types.map(_parseAccess));
+    this._extendedQueries = this._extendedQueries.concat(queries.map(_parseAccess));
+    this._extendedMutations = this._extendedMutations.concat(mutations.map(_parseAccess));
   }
 
   /**
@@ -257,7 +293,7 @@ module.exports = class Keystone {
     ).then(() => {});
   }
 
-  getAdminMeta() {
+  getAdminMeta({ schemaName }) {
     // We've consciously made a design choice that the `read` permission on a
     // list is a master switch in the Admin UI (not the GraphQL API).
     // Justification: If you want to Create without the Read permission, you
@@ -270,18 +306,27 @@ module.exports = class Keystone {
     // In all these cases, the Admin UI becomes unnecessarily complex.
     // So we only allow all these actions if you also have read access.
     const lists = arrayToObject(
-      this.listsArray.filter(list => list.access.read && !list.isAuxList),
+      this.listsArray.filter(list => list.access[schemaName].read && !list.isAuxList),
       'key',
-      list => list.getAdminMeta()
+      list => list.getAdminMeta({ schemaName })
     );
 
     return { lists, name: this.name };
   }
 
-  getTypeDefs({ skipAccessControl = false } = {}) {
+  getTypeDefs({ schemaName }) {
     // Aux lists are only there for typing and internal operations, they should
     // not have any GraphQL operations performed on them
     const firstClassLists = this.listsArray.filter(list => !list.isAuxList);
+
+    const mutations = unique(
+      flatten([
+        ...firstClassLists.map(list => list.getGqlMutations({ schemaName })),
+        this._extendedMutations
+          .filter(({ access }) => access[schemaName])
+          .map(({ schema }) => schema),
+      ])
+    );
 
     // Fields can be represented multiple times within and between lists.
     // If a field defines a `getGqlAuxTypes()` method, it will be
@@ -289,8 +334,10 @@ module.exports = class Keystone {
     // graphql-tools will blow up (rightly so) on duplicated types.
     // Deduping here avoids that problem.
     return [
-      ...unique(flatten(this.listsArray.map(list => list.getGqlTypes({ skipAccessControl })))),
-      ...unique(this._extendedTypes),
+      ...unique(flatten(this.listsArray.map(list => list.getGqlTypes({ schemaName })))),
+      ...unique(
+        this._extendedTypes.filter(({ access }) => access[schemaName]).map(({ type }) => type)
+      ),
       `"""NOTE: Can be JSON, or a Boolean/Int/String
           Why not a union? GraphQL doesn't support a union including a scalar
           (https://github.com/facebook/graphql/issues/215)"""
@@ -314,7 +361,11 @@ module.exports = class Keystone {
           """Access Control settings for the currently logged in (or anonymous)
              user when performing 'delete' operations."""
           delete: JSON
-       }`,
+
+          """Access Control settings for the currently logged in (or anonymous)
+             user when performing 'auth' operations."""
+          auth: JSON
+        }`,
       `type _ListSchemaRelatedFields {
         """The typename as used in GraphQL queries"""
         type: String
@@ -351,22 +402,22 @@ module.exports = class Keystone {
       `type Query {
           ${unique(
             flatten([
-              ...firstClassLists.map(list => list.getGqlQueries({ skipAccessControl })),
-              this._extendedQueries.map(({ schema }) => schema),
+              ...firstClassLists.map(list => list.getGqlQueries({ schemaName })),
+              this._extendedQueries
+                .filter(({ access }) => access[schemaName])
+                .map(({ schema }) => schema),
             ])
           ).join('\n')}
           """ Retrieve the meta-data for all lists. """
           _ksListsMeta: [_ListMeta]
        }`,
-      `type Mutation {
-          ${unique(
-            flatten([
-              ...firstClassLists.map(list => list.getGqlMutations({ skipAccessControl })),
-              this._extendedMutations.map(({ schema }) => schema),
-            ])
-          ).join('\n')}
+      mutations.length > 0 &&
+        `type Mutation {
+          ${mutations.join('\n')}
        }`,
-    ].map(s => print(gql(s)));
+    ]
+      .filter(s => s)
+      .map(s => print(gql(s)));
   }
 
   // It's not Keystone core's responsibility to create an executable schema, but
@@ -377,8 +428,8 @@ module.exports = class Keystone {
       graphql(schema, query, null, context, variables);
   }
 
-  getAdminSchema() {
-    const typeDefs = this.getTypeDefs();
+  getAdminSchema({ schemaName }) {
+    const typeDefs = this.getTypeDefs({ schemaName });
     if (debugGraphQLSchemas()) {
       typeDefs.forEach(i => console.log(i));
     }
@@ -401,6 +452,7 @@ module.exports = class Keystone {
       read: access => access.getRead(),
       update: access => access.getUpdate(),
       delete: access => access.getDelete(),
+      auth: access => access.getAuth(),
     };
 
     // Aux lists are only there for typing and internal operations, they should
@@ -422,8 +474,8 @@ module.exports = class Keystone {
             fields: flatten(
               list
                 .getFieldsRelatedTo(key)
-                .filter(field => field.access.read)
-                .map(field => Object.keys(field.gqlOutputFieldResolvers))
+                .filter(field => field.access[schemaName].read)
+                .map(field => Object.keys(field.gqlOutputFieldResolvers({ schemaName })))
             ),
           }))
           .filter(({ fields }) => fields.length),
@@ -435,43 +487,80 @@ module.exports = class Keystone {
     // TODO: Document this order of precendence, because it's not obvious, and
     // there's no errors thrown
     // TODO: console.warn when duplicate keys are detected?
-    const customResolver = ({ schema, resolver }) => {
-      const name = gql(`type t { ${schema} }`).definitions[0].fields[0].name.value;
+    const customResolver = type => ({ schema, resolver, access }) => {
+      const gqlName = gql(`type t { ${schema} }`).definitions[0].fields[0].name.value;
+
+      // Perform access control check before passing off control to the
+      // user defined resolver (along with the evalutated access).
+      const computeAccess = context => {
+        const _access = context.getCustomAccessControlForUser(access);
+        if (!_access) {
+          graphqlLogger.debug({ access, gqlName }, 'Access statically or implicitly denied');
+          graphqlLogger.info({ gqlName }, 'Access Denied');
+          // If the client handles errors correctly, it should be able to
+          // receive partial data (for the fields the user has access to),
+          // and then an `errors` array of AccessDeniedError's
+          throw new AccessDeniedError({
+            data: { type, target: gqlName },
+            internalData: {
+              authedId: context.authedItem && context.authedItem.id,
+              authedListKey: context.authedListKey,
+            },
+          });
+        }
+        return _access;
+      };
       return {
-        [name]: (obj, args, context, info) =>
-          resolver(obj, args, context, info, { query: this._buildQueryHelper(context) }),
+        [gqlName]: (obj, args, context, info) =>
+          resolver(obj, args, context, info, {
+            query: this._buildQueryHelper(context),
+            access: computeAccess(context),
+          }),
       };
     };
-    const resolvers = {
-      // Order of spreading is important here - we don't want user-defined types
-      // to accidentally override important things like `Query`.
-      ...objMerge(this.listsArray.map(list => list.gqlAuxFieldResolvers)),
-      ...objMerge(this.listsArray.map(list => list.gqlFieldResolvers)),
+    const resolvers = filterValues(
+      {
+        // Order of spreading is important here - we don't want user-defined types
+        // to accidentally override important things like `Query`.
+        ...objMerge(this.listsArray.map(list => list.gqlAuxFieldResolvers({ schemaName }))),
+        ...objMerge(this.listsArray.map(list => list.gqlFieldResolvers({ schemaName }))),
 
-      JSON: GraphQLJSON,
+        JSON: GraphQLJSON,
 
-      _QueryMeta: queryMetaResolver,
-      _ListMeta: listMetaResolver,
-      _ListAccess: listAccessResolver,
-      _ListSchema: listSchemaResolver,
+        _QueryMeta: queryMetaResolver,
+        _ListMeta: listMetaResolver,
+        _ListAccess: listAccessResolver,
+        _ListSchema: listSchemaResolver,
 
-      Query: {
-        // Order is also important here, any TypeQuery's defined by types
-        // shouldn't be able to override list-level queries
-        ...objMerge(firstClassLists.map(list => list.gqlAuxQueryResolvers)),
-        ...objMerge(firstClassLists.map(list => list.gqlQueryResolvers)),
-        // And the Keystone meta queries must always be available
-        _ksListsMeta: (_, args, context) =>
-          this.listsArray.filter(list => list.access.read).map(list => list.listMeta(context)),
-        ...objMerge(this._extendedQueries.map(customResolver)),
+        Query: {
+          // Order is also important here, any TypeQuery's defined by types
+          // shouldn't be able to override list-level queries
+          ...objMerge(firstClassLists.map(list => list.gqlAuxQueryResolvers())),
+          ...objMerge(firstClassLists.map(list => list.gqlQueryResolvers({ schemaName }))),
+          // And the Keystone meta queries must always be available
+          _ksListsMeta: (_, args, context) =>
+            this.listsArray
+              .filter(list => list.access[schemaName].read)
+              .map(list => list.listMeta(context)),
+          ...objMerge(
+            this._extendedQueries
+              .filter(({ access }) => access[schemaName])
+              .map(customResolver('query'))
+          ),
+        },
+
+        Mutation: {
+          ...objMerge(firstClassLists.map(list => list.gqlAuxMutationResolvers())),
+          ...objMerge(firstClassLists.map(list => list.gqlMutationResolvers({ schemaName }))),
+          ...objMerge(
+            this._extendedMutations
+              .filter(({ access }) => access[schemaName])
+              .map(customResolver('mutation'))
+          ),
+        },
       },
-
-      Mutation: {
-        ...objMerge(firstClassLists.map(list => list.gqlAuxMutationResolvers)),
-        ...objMerge(firstClassLists.map(list => list.gqlMutationResolvers)),
-        ...objMerge(this._extendedMutations.map(customResolver)),
-      },
-    };
+      o => Object.entries(o).length > 0
+    );
 
     if (debugGraphQLSchemas()) {
       console.log(resolvers);
@@ -488,14 +577,14 @@ module.exports = class Keystone {
     };
   }
 
-  dumpSchema(file) {
+  dumpSchema(file, schemaName) {
     // The 'Upload' scalar is normally automagically added by Apollo Server
     // See: https://blog.apollographql.com/file-uploads-with-apollo-server-2-0-5db2f3f60675
     // Since we don't execute apollo server over this schema, we have to
     // reinsert it.
     const schema = `
       scalar Upload
-      ${this.getTypeDefs({ skipAccessControl: true }).join('\n')}
+      ${this.getTypeDefs({ schemaName }).join('\n')}
     `;
     fs.writeFileSync(file, schema);
   }
@@ -534,12 +623,26 @@ module.exports = class Keystone {
     return mergeRelationships(createdItems, createdRelationships);
   }
 
-  async prepare({ dev = false, apps = [], distDir } = {}) {
+  async prepare({
+    dev = false,
+    apps = [],
+    distDir,
+    pinoOptions,
+    cors = { origin: true, credentials: true },
+  } = {}) {
     const middlewares = flattenDeep([
       // Used by other middlewares such as authentication strategies. Important
       // to be first so the methods added to `req` are available further down
       // the request pipeline.
-      commonSessionMiddleware(this, this._cookieSecret, this._sessionStore),
+      commonSessionMiddleware({
+        keystone: this,
+        cookieSecret: this._cookieSecret,
+        sessionStore: this.sessionStore,
+        secureCookies: this._secureCookies,
+        cookieMaxAge: this._cookieMaxAge,
+      }),
+      falsey(process.env.DISABLE_LOGGING) && require('express-pino-logger')(pinoOptions),
+      cors && createCorsMiddleware(cors),
       ...(await Promise.all(
         [
           // Inject any field middlewares (eg; WYSIWIG's static assets)