+++ packages/keystone/lib/List/index.js
@@ -139,6 +139,7 @@ module.exports = class List {
       registerType,
       createAuxList,
       isAuxList,
+      schemaNames,
     }
   ) {
     this.key = key;
@@ -213,8 +214,10 @@ module.exports = class List {
 
     this.adapterName = adapter.name;
     this.adapter = adapter.newListAdapter(this.key, adapterConfig);
+    this._schemaNames = schemaNames;
 
     this.access = parseListAccess({
+      schemaNames: this._schemaNames,
       listKey: key,
       access,
       defaultAccess: this.defaultAccess.list,
@@ -297,6 +300,8 @@ module.exports = class List {
           fieldAdapterClass: type.adapters[this.adapterName],
           defaultAccess: this.defaultAccess.field,
           createAuxList: this.createAuxList,
+          schemaNames: this._schemaNames,
+          listAccess: this.access,
         })
     );
     this.fields = Object.values(this.fieldsByPath);
@@ -305,18 +310,21 @@ module.exports = class List {
     );
   }
 
-  getAdminMeta() {
+  getAdminMeta({ schemaName }) {
+    const schemaAccess = this.access[schemaName];
     return {
       key: this.key,
       // Reduce to truthy values (functions can't be passed over the webpack
       // boundary)
-      access: mapKeys(this.access, val => !!val),
+      access: mapKeys(schemaAccess, val => !!val),
       label: this.adminUILabels.label,
       singular: this.adminUILabels.singular,
       plural: this.adminUILabels.plural,
       path: this.adminUILabels.path,
       gqlNames: this.gqlNames,
-      fields: this.fields.filter(field => field.access.read).map(field => field.getAdminMeta()),
+      fields: this.fields
+        .filter(field => field.access[schemaName].read)
+        .map(field => field.getAdminMeta({ schemaName })),
       views: this.views,
       adminConfig: {
         defaultPageSize: this.adminConfig.defaultPageSize,
@@ -330,18 +338,19 @@ module.exports = class List {
     };
   }
 
-  getGqlTypes({ skipAccessControl = false } = {}) {
+  getGqlTypes({ schemaName }) {
+    const schemaAccess = this.access[schemaName];
     // https://github.com/opencrud/opencrud/blob/master/spec/2-relational/2-2-queries/2-2-3-filters.md#boolean-expressions
     const types = [];
     if (
-      skipAccessControl ||
-      this.access.read ||
-      this.access.create ||
-      this.access.update ||
-      this.access.delete
+      schemaAccess.read ||
+      schemaAccess.create ||
+      schemaAccess.update ||
+      schemaAccess.delete ||
+      schemaAccess.auth
     ) {
       types.push(
-        ...flatten(this.fields.map(field => field.getGqlAuxTypes({ skipAccessControl }))),
+        ...flatten(this.fields.map(field => field.getGqlAuxTypes({ schemaName }))),
         `
         """ ${this.schemaDoc || 'A keystone list'} """
         type ${this.gqlNames.outputTypeName} {
@@ -355,11 +364,11 @@ module.exports = class List {
           _label_: String
           ${flatten(
             this.fields
-              .filter(field => skipAccessControl || field.access.read) // If it's globally set to false, makes sense to never show it
+              .filter(field => field.access[schemaName].read) // If it's globally set to false, makes sense to never show it
               .map(field =>
                 field.schemaDoc
-                  ? `""" ${field.schemaDoc} """ ${field.gqlOutputFields}`
-                  : field.gqlOutputFields
+                  ? `""" ${field.schemaDoc} """ ${field.gqlOutputFields({ schemaName })}`
+                  : field.gqlOutputFields({ schemaName })
               )
           ).join('\n')}
         }
@@ -371,8 +380,8 @@ module.exports = class List {
 
           ${flatten(
             this.fields
-              .filter(field => skipAccessControl || field.access.read) // If it's globally set to false, makes sense to never show it
-              .map(field => field.gqlQueryInputFields)
+              .filter(field => field.access[schemaName].read) // If it's globally set to false, makes sense to never show it
+              .map(field => field.gqlQueryInputFields({ schemaName }))
           ).join('\n')}
         }`,
         // TODO: Include other `unique` fields and allow filtering by them
@@ -383,13 +392,13 @@ module.exports = class List {
       );
     }
 
-    if (skipAccessControl || this.access.update) {
+    if (schemaAccess.update) {
       types.push(`
         input ${this.gqlNames.updateInputName} {
           ${flatten(
             this.fields
               .filter(({ path }) => path !== 'id') // Exclude the id fields update types
-              .filter(field => skipAccessControl || field.access.update) // If it's globally set to false, makes sense to never let it be updated
+              .filter(field => field.access[schemaName].update) // If it's globally set to false, makes sense to never let it be updated
               .map(field => field.gqlUpdateInputFields)
           ).join('\n')}
         }
@@ -402,13 +411,13 @@ module.exports = class List {
       `);
     }
 
-    if (skipAccessControl || this.access.create) {
+    if (schemaAccess.create) {
       types.push(`
         input ${this.gqlNames.createInputName} {
           ${flatten(
             this.fields
               .filter(({ path }) => path !== 'id') // Exclude the id fields create types
-              .filter(field => skipAccessControl || field.access.create) // If it's globally set to false, makes sense to never let it be created
+              .filter(field => field.access[schemaName].create) // If it's globally set to false, makes sense to never let it be created
               .map(field => field.gqlCreateInputFields)
           ).join('\n')}
         }
@@ -420,7 +429,7 @@ module.exports = class List {
       `);
     }
 
-    if (this.hasAuth()) {
+    if (this.hasAuth() && schemaAccess.auth) {
       // If auth is enabled for this list (doesn't matter what strategy)
       types.push(`
         type ${this.gqlNames.unauthenticateOutputName} {
@@ -455,15 +464,14 @@ module.exports = class List {
     ];
   }
 
-  getGqlQueries({ skipAccessControl = false } = {}) {
+  getGqlQueries({ schemaName }) {
+    const schemaAccess = this.access[schemaName];
     // All the auxiliary queries the fields want to add
-    const queries = flatten(
-      this.fields.map(field => field.getGqlAuxQueries({ skipAccessControl }))
-    );
+    const queries = flatten(this.fields.map(field => field.getGqlAuxQueries()));
 
     // If `read` is either `true`, or a function (we don't care what the result
     // of the function is, that'll get executed at a later time)
-    if (skipAccessControl || this.access.read) {
+    if (schemaAccess.read) {
       queries.push(
         `
         """ Search for all ${this.gqlNames.outputTypeName} items which match the where clause. """
@@ -491,7 +499,7 @@ module.exports = class List {
       );
     }
 
-    if (this.hasAuth()) {
+    if (this.hasAuth() && schemaAccess.auth) {
       // If auth is enabled for this list (doesn't matter what strategy)
       queries.push(`${this.gqlNames.authenticatedQueryName}: ${this.gqlNames.outputTypeName}`);
     }
@@ -544,8 +552,9 @@ module.exports = class List {
     };
   }
 
-  get gqlFieldResolvers() {
-    if (!this.access.read) {
+  gqlFieldResolvers({ schemaName }) {
+    const schemaAccess = this.access[schemaName];
+    if (!schemaAccess.read) {
       return {};
     }
     const fieldResolvers = {
@@ -553,10 +562,10 @@ module.exports = class List {
       _label_: this.labelResolver,
       ...objMerge(
         this.fields
-          .filter(field => field.access.read)
+          .filter(field => field.access[schemaName].read)
           .map(field =>
             // Get the resolvers for the (possibly multiple) output fields and wrap each with access control
-            mapKeys(field.gqlOutputFieldResolvers, innerResolver =>
+            mapKeys(field.gqlOutputFieldResolvers({ schemaName }), innerResolver =>
               this.wrapFieldResolverWithAC(field, innerResolver)
             )
           )
@@ -565,29 +574,37 @@ module.exports = class List {
     return { [this.gqlNames.outputTypeName]: fieldResolvers };
   }
 
-  get gqlAuxFieldResolvers() {
-    // TODO: Obey the same ACL rules based on parent type
-    return objMerge(this.fields.map(field => field.gqlAuxFieldResolvers));
+  gqlAuxFieldResolvers({ schemaName }) {
+    const schemaAccess = this.access[schemaName];
+    if (
+      schemaAccess.read ||
+      schemaAccess.create ||
+      schemaAccess.update ||
+      schemaAccess.delete ||
+      schemaAccess.auth
+    ) {
+      return objMerge(this.fields.map(field => field.gqlAuxFieldResolvers({ schemaName })));
+    }
+    return {};
   }
 
-  get gqlAuxQueryResolvers() {
+  gqlAuxQueryResolvers() {
     // TODO: Obey the same ACL rules based on parent type
-    return objMerge(this.fields.map(field => field.gqlAuxQueryResolvers));
+    return objMerge(this.fields.map(field => field.gqlAuxQueryResolvers()));
   }
 
-  get gqlAuxMutationResolvers() {
+  gqlAuxMutationResolvers() {
     // TODO: Obey the same ACL rules based on parent type
-    return objMerge(this.fields.map(field => field.gqlAuxMutationResolvers));
+    return objMerge(this.fields.map(field => field.gqlAuxMutationResolvers()));
   }
 
-  getGqlMutations({ skipAccessControl = false } = {}) {
-    const mutations = flatten(
-      this.fields.map(field => field.getGqlAuxMutations({ skipAccessControl }))
-    );
+  getGqlMutations({ schemaName }) {
+    const schemaAccess = this.access[schemaName];
+    const mutations = flatten(this.fields.map(field => field.getGqlAuxMutations()));
 
     // NOTE: We only check for truthy as it could be `true`, or a function (the
     // function is executed later in the resolver)
-    if (skipAccessControl || this.access.create) {
+    if (schemaAccess.create) {
       mutations.push(`
         """ Create a single ${this.gqlNames.outputTypeName} item. """
         ${this.gqlNames.createMutationName}(
@@ -603,7 +620,7 @@ module.exports = class List {
       `);
     }
 
-    if (skipAccessControl || this.access.update) {
+    if (schemaAccess.update) {
       mutations.push(`
       """ Update a single ${this.gqlNames.outputTypeName} item by ID. """
         ${this.gqlNames.updateMutationName}(
@@ -620,7 +637,7 @@ module.exports = class List {
       `);
     }
 
-    if (skipAccessControl || this.access.delete) {
+    if (schemaAccess.delete) {
       mutations.push(`
         """ Delete a single ${this.gqlNames.outputTypeName} item by ID. """
         ${this.gqlNames.deleteMutationName}(
@@ -636,7 +653,7 @@ module.exports = class List {
       `);
     }
 
-    if (this.hasAuth()) {
+    if (this.hasAuth() && schemaAccess.auth) {
       // If auth is enabled for this list (doesn't matter what strategy)
       mutations.push(
         `${this.gqlNames.unauthenticateMutationName}: ${this.gqlNames.unauthenticateOutputName}`
@@ -821,12 +838,13 @@ module.exports = class List {
     return await this._itemsQuery({ where: { ...remainingAccess, ...idFilters } });
   }
 
-  get gqlQueryResolvers() {
+  gqlQueryResolvers({ schemaName }) {
+    const schemaAccess = this.access[schemaName];
     let resolvers = {};
 
     // If set to false, we can confidently remove these resolvers entirely from
     // the graphql schema
-    if (this.access.read) {
+    if (schemaAccess.read) {
       resolvers = {
         [this.gqlNames.listQueryName]: (_, args, context) =>
           this.listQuery(args, context, this.gqlNames.listQueryName),
@@ -844,7 +862,7 @@ module.exports = class List {
     // NOTE: This query is not effected by the read permissions; if the user can
     // authenticate themselves, then they already have access to know that the
     // list exists
-    if (this.hasAuth()) {
+    if (this.hasAuth() && schemaAccess.auth) {
       resolvers[this.gqlNames.authenticatedQueryName] = (_, __, context) =>
         this.authenticatedQuery(context);
     }
@@ -886,6 +904,7 @@ module.exports = class List {
         getRead: () => context.getListAccessControlForUser(this.key, undefined, 'read'),
         getUpdate: () => context.getListAccessControlForUser(this.key, undefined, 'update'),
         getDelete: () => context.getListAccessControlForUser(this.key, undefined, 'delete'),
+        getAuth: () => context.getListAccessControlForUser(this.key, undefined, 'auth'),
       }),
       getSchema: () => {
         const queries = [
@@ -952,10 +971,16 @@ module.exports = class List {
     if (first < Infinity && first > maxResults) {
       throwLimitsExceeded({ type: 'maxResults', limit: maxResults });
     }
-    // + 1 to allow limit violation detection
-    const resultsLimit = Math.min(maxResults + 1, first);
-    if (resultsLimit < Infinity) {
-      args.first = resultsLimit;
+    if (!(extra && extra.meta)) {
+      // "first" is designed to truncate the count value, but accurate counts are still
+      // needed for pagination.  resultsLimit is meant for protecting KS memory usage,
+      // not DB performance, anyway, so resultsLimit is only applied to queries that
+      // could return many results.
+      // + 1 to allow limit violation detection
+      const resultsLimit = Math.min(maxResults + 1, first);
+      if (resultsLimit < Infinity) {
+        args.first = resultsLimit;
+      }
     }
     const results = await this.adapter.itemsQuery(args, extra);
     if (results.length > maxResults) {
@@ -969,14 +994,19 @@ module.exports = class List {
       return null;
     }
 
+    const queryName = this.gqlNames.authenticatedQueryName;
+    const access = this.checkListAccess(context, undefined, 'auth', { queryName });
     return this.itemQuery(
-      { where: { id: context.authedItem.id } },
+      mergeWhereClause({ where: { id: context.authedItem.id } }, access),
       context,
       this.gqlNames.authenticatedQueryName
     );
   }
 
   async authenticateMutation(authType, args, context) {
+    const queryName = getAuthMutationName(this.gqlNames.authenticateMutationPrefix, authType);
+    this.checkListAccess(context, undefined, 'auth', { queryName });
+
     // This is currently hard coded to enable authenticating with the admin UI.
     // In the near future we will set up the admin-ui application and api to be
     // non-public.
@@ -999,14 +1029,18 @@ module.exports = class List {
   }
 
   async unauthenticateMutation(context) {
+    const queryName = this.gqlNames.unauthenticateMutationName;
+    this.checkListAccess(context, undefined, 'auth', { queryName });
+
     await context.endAuthedSession();
     return { success: true };
   }
 
-  get gqlMutationResolvers() {
+  gqlMutationResolvers({ schemaName }) {
+    const schemaAccess = this.access[schemaName];
     const mutationResolvers = {};
 
-    if (this.access.create) {
+    if (schemaAccess.create) {
       mutationResolvers[this.gqlNames.createMutationName] = (_, { data }, context) =>
         this.createMutation(data, context);
 
@@ -1014,7 +1048,7 @@ module.exports = class List {
         this.createManyMutation(data, context);
     }
 
-    if (this.access.update) {
+    if (schemaAccess.update) {
       mutationResolvers[this.gqlNames.updateMutationName] = (_, { id, data }, context) =>
         this.updateMutation(id, data, context);
 
@@ -1022,7 +1056,7 @@ module.exports = class List {
         this.updateManyMutation(data, context);
     }
 
-    if (this.access.delete) {
+    if (schemaAccess.delete) {
       mutationResolvers[this.gqlNames.deleteMutationName] = (_, { id }, context) =>
         this.deleteMutation(id, context);
 
@@ -1033,7 +1067,7 @@ module.exports = class List {
     // NOTE: This query is not effected by the read permissions; if the user can
     // authenticate themselves, then they already have access to know that the
     // list exists
-    if (this.hasAuth()) {
+    if (this.hasAuth() && schemaAccess.auth) {
       mutationResolvers[this.gqlNames.unauthenticateMutationName] = (_, __, context) =>
         this.unauthenticateMutation(context);